move health checker to utils for reuse
